+++ packages/cli/src/__tests__/make-ceramic-core.ts
@@ -1,12 +1,10 @@
 import { IpfsApi } from '@ceramicnetwork/common'
 import { Ceramic } from '@ceramicnetwork/core'
 import { TileDocumentHandler } from '@ceramicnetwork/stream-tile-handler'
-import { StreamID } from '@ceramicnetwork/streamid'
 
 export async function makeCeramicCore(
   ipfs: IpfsApi,
-  stateStoreDirectory: string,
-  modelsToIndex: Array<string> = []
+  stateStoreDirectory: string
 ): Promise<Ceramic> {
   const core = await Ceramic.create(ipfs, {
     adminDids: ['did:key:z6MkgwMzPmLuvUiWsQfyQeGpNRRNkLyCB5cL96fPshy1DKJd'],
@@ -15,7 +13,7 @@ export async function makeCeramicCore(
     anchorOnRequest: false,
     indexing: {
       db: `sqlite://${stateStoreDirectory}/ceramic.sqlite`,
-      models: modelsToIndex.map(StreamID.fromString),
+      models: [],
       allowQueriesBeforeHistoricalSync: true,
     },
   })