+++ packages/http-client/src/remote-index-api.ts
@@ -6,6 +6,7 @@ import type {
   StreamState,
 } from '@ceramicnetwork/common'
 import { StreamUtils, fetchJson } from '@ceramicnetwork/common'
+import { serializeObjectToSearchParams } from './utils.js'
 
 /**
  * IndexAPI implementation on top of HTTP endpoint.
@@ -14,23 +15,25 @@ export class RemoteIndexApi implements IndexApi {
   // Stored as a member to make it easier to inject a mock in unit tests
   private readonly _fetchJson: typeof fetchJson = fetchJson
   private readonly _collectionURL: URL
+  private readonly _countURL: URL
 
   constructor(apiUrl: URL) {
     this._collectionURL = new URL('./collection', apiUrl)
+    this._countURL = new URL('./collection/count', apiUrl)
   }
 
   async count(query: BaseQuery): Promise<number> {
-    throw new Error('Not implemented')
+    const queryURL = serializeObjectToSearchParams(this._countURL, query)
+    const response = await this._fetchJson(queryURL)
+    return response.count
   }
 
   /**
    * Issue a query to `/collection` endpoint.
    */
   async query(query: PaginationQuery): Promise<Page<StreamState | null>> {
-    const queryURL = new URL(this._collectionURL)
-    for (const key in query) {
-      queryURL.searchParams.set(key, query[key])
-    }
+    const queryURL = serializeObjectToSearchParams(this._collectionURL, query)
+
     const response = await this._fetchJson(queryURL)
     const edges = response.edges.map((e) => {
       return {